pyproject.toml
[tool.poetry]
name = "poc-agent"
version = "0.1.0"
description = "Agentic POC for JSON source-sink conversion"
authors = ["Swetha"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.110"
uvicorn = "^0.27"
langchain = "^0.2"
langchain-google-genai = "^1.0"
google-generativeai = "^0.5"
sqlalchemy = "^2.0"
psycopg2-binary = "^2.9"
python-dotenv = "^1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
